Fix functions that should be inline in *.templates.h files
